+++ _Frontends/UI/k45_adr_vuio/src/components/StationBuildingOptions.tsx
@@ -1,17 +1,17 @@
 import { Entity, SelectInfoPanelService, SelectedInfoOptions, nameToString, replaceArgs } from "@klyte45/adr-commons";
 import { VanillaComponentResolver } from '@klyte45/vuio-commons';
-import { Component, useState } from "react";
+import { useState } from "react";
 import { translate } from "utility/translate";
 
-type Props = { entity: Entity, response: SelectedInfoOptions, onChanged: () => Promise<any> };
+type Props = { entityOrigin: Entity, response: SelectedInfoOptions, onChanged: () => Promise<any> };
 
-export const StationBuildingOptionsComponent = ({ entity, onChanged, response }: Props) => {
+export const StationBuildingOptionsComponent = ({ entityOrigin, onChanged, response }: Props) => {
 
     const [loading, setLoading] = useState(false);
 
     const setReference = async (target: Entity, reference: Entity) => {
         setLoading(true);
-        await SelectInfoPanelService.setEntityNamingRef(target, reference);
+        await SelectInfoPanelService.setEntityNamingRef(target, reference, entityOrigin);
         await onChanged();
         setLoading(false);
     }
@@ -24,14 +24,14 @@ export const StationBuildingOptionsComponent = ({ entity, onChanged, response }:
     const currentVal = response.entityValue;
     const focusKey = VR.FOCUS_DISABLED;
     return <>
-        <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(entity, response.districtRef)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={isDistrictSelected} />{translate("StationBuildingOptions.UseDistrictOpt")}</>} />
+        <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(response.targetEntityToName, response.districtRef)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={isDistrictSelected} />{translate("StationBuildingOptions.UseDistrictOpt")}</>} />
         {
             response.roadAggegateOptions.map(x =>
-                <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(entity, x.entity)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={currentVal.Index == x.entity.Index} />{replaceArgs(translate("StationBuildingOptions.RoadOptPattern", "{name}"), { name: nameToString(x.name)!! })}</>} />)
+                <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(response.targetEntityToName, x.entity)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={currentVal.Index == x.entity.Index} />{replaceArgs(translate("StationBuildingOptions.RoadOptPattern", "{name}"), { name: nameToString(x.name)!! })}</>} />)
         }
         {
             response.buildingsOptions.map(x =>
-                <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(entity, x.entity)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={currentVal.Index == x.entity.Index} />{replaceArgs(translate("StationBuildingOptions.BuildingOptPattern", "{name}"), { name: nameToString(x.name)!! })}</>} />)
+                <VR.InfoRow subRow={true} className={rowTheme} left={<><VR.RadioToggle onChange={() => setReference(response.targetEntityToName, x.entity)} focusKey={focusKey} className={VR.themeToggleLine.toggle} checked={currentVal.Index == x.entity.Index} />{replaceArgs(translate("StationBuildingOptions.BuildingOptPattern", "{name}"), { name: nameToString(x.name)!! })}</>} />)
         }
     </>