+++ BelzontAdr/changelog.md
@@ -2,3 +2,4 @@
 - Removed feature to autoname based on districts - use manual selection instead!
 - Merged Districts and Road override settings
 - Fixed behavior of naming based on districts when the vanilla generator is being used
+- Now it shows the building name along the ADR generated name for sub-buildings selected information panel